#181 / Request, Response 생성 및 코드 정리
