+++ Qapple/Qapple/Data/Network/DataMapping/Request/CommentRequest.swift
@@ -1,0 +1,16 @@
+//
+//  CommentRequest.swift
+//  Qapple
+//
+//  Created by 문인범 on 9/4/24.
+//
+
+import Foundation
+
+
+class CommentRequest {
+    struct UploadComment {
+        let boardId: Int
+        let content: String
+    }
+}
